requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.3
sentence-transformers>=2.2.0
chromadb>=0.4.0
pypdf>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0


